Moved app initialization outside of index.js
